+++ examples/todos/src/index.js
@@ -2,15 +2,13 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { AtomRoot } from 'atom-state'
 import './index.css'
-import Counter from './Counter'
+import App from './App'
 import store from './store'
-import CountInfo from './CountInfo'
 
 ReactDOM.render(
   <React.StrictMode>
     <AtomRoot store={store}>
-      <Counter />
-      <CountInfo />
+      <App />
     </AtomRoot>
   </React.StrictMode>,
   document.getElementById('root')